move useStateAccessMap
